Project/Server/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.3.4; python_version >= '3.6'
certifi==2020.12.5
distlib==0.3.1
dj-database-url==0.5.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-jwt==0.3.1
django==3.2
filelock==3.0.12
graphene-django==2.15.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
install==1.3.4
pillow==8.1.0
pipenv==2020.11.15
promise==2.3
psycopg2-binary==2.8.6
pyjwt==1.7.0
python-decouple==3.4
python-http-client==3.3.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
rx==1.6.1
sendgrid==6.7.0
singledispatch==3.6.1; python_version >= '2.6'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
starkbank-ecdsa==1.1.0
text-unidecode==1.3
virtualenv-clone==0.5.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
virtualenv==20.4.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
